MailChimp Connect Tweaks
I made some changes here based on observations I saw during our big RG campaign.

We hit the MC API pretty hard to pull the Interests/Categories (Genres) lists, so
I added object-caching around it.  With a stock WP install, this will at least
cache during the page generation, but with back-end cache configured, it will
now cache for much longer.  I've added the 'memcached-is-your-friend' plugin which
adds the memcache or memcached extension as a WP Object Cache provider.  It's very
simple, and in testing seems to work wonderfully.  This should improve our page load
times a bit and reduce the amount of calls we make to MailChimp.

I made a few bug fixes such as needing to add quotes around the RG_INCLUDE_PATH define
which threw warnings, and adding a quote onto the end of a page redirect to the Thank
You Page after updating a subscription.

I also made a tweak to make it easier to update the MailChimp API URL if needed by
defining it up at the top of the file.

I also changed the way we generate our JSON requests by using standard PHP objects
and then json_encode()'ing them, which is easier to code around and reduces the chances
of errors creating JSON by hand.