TTR-151 - RG - signup form widgets for use on 3rd party sites
I created a new Wordpress Widget which also provides a short-code for embedding inside
a regular post using [rg_signup].  The reason I had to make it a widget as well is because
you can't use short-codes inside widgets without hacking the theme or installing another
plug-in.

I added the new  Widget to the plugins directory under plugins/rg-subscribe, though it's not
intended for use on our own site.  This plugin didn't really justify its' own repo either so
this place seemed as good as any.  A Zip copy exists in the assets directory and will need to
be updated if the Plugin itself is ever changed.

E.g. From wp-content/plugins/:

$ zip -r wp-content/plugins/rg-mailchimp-connect/assets/RunawayGoodness_Wordpress_Subscription_Widget.zip wp-content/plugins/rg-subscribe

Now users can be referred to https://runawaygoodness.com/wp-content/plugins/rg-mailchimp-connect/assets/RunawayGoodness_Wordpress_Subscription_Widget.zip
to download a copy of the Wordpress Widget.

https://teamtrope.atlassian.net/browse/TTR-151